# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetConnect API
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned after login or profile completion.
    SetupToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Setup token returned after OTP verification for onboarding.
...
paths:
...
  /uploads/user/avatar:
    post:
      tags: [UserUploads]
      summary: Upload user avatar (accepts setup or access token)
      security:
        - SetupToken: []
        - AccessToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      avatarUrl:
                        type: string
                        format: uri
                      url:
                        type: string
                        format: uri
                      key:
                        type: string
        "400":
          description: Upload failed
        "401":
          description: Unauthorized
  /uploads/user/cover:
    post:
      tags: [UserUploads]
      summary: Upload user cover (accepts setup or access token)
      security:
        - SetupToken: []
        - AccessToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Cover uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      coverUrl:
                        type: string
                        format: uri
                      url:
                        type: string
                        format: uri
                      key:
                        type: string
        "400":
          description: Upload failed
        "401":
          description: Unauthorized
  /uploads/pets/{id}/avatar:
    post:
      tags: [UserUploads]
      summary: Upload pet avatar
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Pet avatar uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      avatarUrl:
                        type: string
                        format: uri
                      url:
                        type: string
                        format: uri
                      key:
                        type: string
        "400":
          description: Upload failed
        "404":
          description: Pet not found
  /uploads/pets/{id}/photos:
    post:
      tags: [UserUploads]
      summary: Upload pet photos
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 1
                  maxItems: 5
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Pet photos uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      photos:
                        type: array
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              format: uri
                            key:
                              type: string
        "400":
          description: Upload failed
        "404":
          description: Pet not found
  /uploads/pets/{id}/documents:
    post:
      tags: [UserUploads]
      summary: Upload pet document
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, file]
              properties:
                title:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Medical document uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      document:
                        type: object
                        properties:
                          title:
                            type: string
                          documentUrl:
                            type: string
                            format: uri
                          uploadedAt:
                            type: string
                            format: date-time
                          key:
                            type: string
        "400":
          description: Upload failed
        "404":
          description: Pet not found
  /uploads/pets/{id}/photos/{fileId}:
    delete:
      tags: [UserUploads]
      summary: Delete pet photo
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Photo deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "404":
          description: Photo not found
  /uploads/pets/{id}/documents/{fileId}:
    delete:
      tags: [UserUploads]
      summary: Delete pet document
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "404":
          description: Document not found
...